0.4 -- MATLAB State Control
Full bi-directional syncing of state and matlab control over the toggle.
Pretty good detection of when animal motion stops (this triggers a
state change).

TODO: Better stop detection
TODO: There is a nagging bug in the transition from state 2 (trial
state) to state (3) wait state. In that, at the onset of the
transitions there is some state flickering (8-11 flips). It always
settles and is hard to see, but it is there. This has to be fixed. I
think it is because the mouse position can stay identical for 8-11
samples and maybe that is throwing off the averaging I am using?