0.5 -- Better Matlab State Control
The flicker bug is fixed. I left some debug code (of course) that was
causing the problem.

Also, I shifted everything to a switch/case block to fit the general
logic better.

Motion detection is pretty good as far as I can tell. This has to be
determined on a comp. mouse and actual mouse basis.