0.99 -- Python with terminal feedback
So the python script is ready for production use. However, there are
some things to be aware of, and some BIG gotchas.

1) I took out a timer for controlling the runs and instead loop until a
certain amount of trials have been run. The variable that controls this
is “trialsToRun.”

2) The data gets saved out as a CSV the end of the run! This means if
you are running the script from a bash shell (or equivalent) and then
interrupt the run, you will be hosed and have no record of anything
(!!!!). So, you need to let the whole thing play out. Alternatively,
you can run the script in an interactive python session (iPython) and
then if you abort the script, the lists that contain all the data will
still be in memory and you can salvage the run.

3) You can get text feedback of the trial and outcome (hit/miss) by
turning on the “giveTerminalFeedback” flag.

**** To Do, Etc. ****
I am working on implementing a pickle stream, which should fix 2.
I am working on implementing some form of real time(ish) plotting.