0.994 -- Basic plot feedback in python
Plot works, but it is a bit limited …
Currently plots time vs. mouse position along with the switch and
report ranges. Each trial is appended to the next, so you are forced to
see them all. The appending is a cheater’s append in that I just clear
and replot.